Bara för att det är nice

+++ 1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/EnemyObject.cpp
@@ -69,7 +69,7 @@ bool EnemyObject::GetType(){
 	return true;
 };
 
-int EnemyObject::Damaged(int playerdmg){
+float EnemyObject::Damaged(float playerdmg){
 	m_hp-=playerdmg;
 	return m_hp;
 }
