+++ 1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/EnemyObject.h
@@ -12,11 +12,11 @@ public:
 	~EnemyObject();
 	bool Update(float deltatime);
 	bool GetType();
-	int Damaged(int playerdmg);
+	float Damaged(float playerdmg);
 	float GetDamage();
 
 protected:
 	float m_cooldown;
-	int m_hp;
+	float m_hp;
 	float m_damage;
 };
